Add counters support
We creates multiple "arenas" to be used by each forked process so they
can do their own locking. Each arena has an a/b set of counters. When a
reader has the lock for a or b, and the writer attempts to update the same
set, it can flip to the other set and avoid waiting for the reader to complete.

We add a simple "stat" command to mdrd to see current counters. It should only
be used for testing; should it die while holding the counter read lock, it
_could_ deadlock mdrd.